# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scores
    description: Pixel Stack global leaderboard
paths:
  /runs:
    post:
      operationId: startRun
      tags: [scores]
      summary: Issue a server-verified game run
      responses:
        "201":
          description: Run issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameRun"
        "429":
          description: Too many run requests
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List the global leaderboard
      responses:
        "200":
          description: The ten highest scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HighScore"
    post:
      operationId: submitScore
      tags: [scores]
      summary: Submit a qualifying score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreInput"
      responses:
        "201":
          description: Score accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HighScore"
        "400":
          description: Invalid score
        "401":
          description: Invalid or expired run proof
        "409":
          description: Score did not qualify
        "429":
          description: Too many score submissions
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    HighScore:
      type: object
      required: [id, nickname, score, created_at]
      properties:
        id:
          type: integer
        nickname:
          type: string
          maxLength: 10
        score:
          type: integer
          minimum: 0
        created_at:
          type: string
    GameRun:
      type: object
      additionalProperties: false
      required: [proof, seed]
      properties:
        proof:
          type: string
          minLength: 1
        seed:
          type: integer
          minimum: 1
          maximum: 4294967295
    RunAction:
      type: object
      additionalProperties: false
      required: [kind, pieceIndex, atMs]
      properties:
        kind:
          type: string
          enum: [anchor]
        pieceIndex:
          type: integer
          minimum: 0
        rotation:
          type: integer
          minimum: 0
          maximum: 3
        col:
          type: integer
          minimum: 0
          maximum: 9
        row:
          type: integer
          minimum: 0
          maximum: 17
        atMs:
          type: integer
          minimum: 0
          maximum: 7200000
    ScoreInput:
      type: object
      additionalProperties: false
      required: [nickname, score, proof, actions, endedAtMs]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 10
          pattern: "^[A-Za-z0-9_-]+$"
        score:
          type: integer
          minimum: 0
        proof:
          type: string
          minLength: 1
        actions:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/RunAction"
        endedAtMs:
          type: integer
          minimum: 0
          maximum: 7200000