# lib/api-zod/src/generated/types/gameRun.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameRun {
  /** @minLength 1 */
  proof: string;
  /**
     * @minimum 1
     * @maximum 4294967295
     */
  seed: number;
}