# lib/api-zod/src/generated/types/runAction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunActionKind } from './runActionKind';

export interface RunAction {
  kind: RunActionKind;
  /** @minimum 0 */
  pieceIndex: number;
  /**
     * @minimum 0
     * @maximum 3
     */
  rotation?: number;
  /**
     * @minimum 0
     * @maximum 9
     */
  col?: number;
  /**
     * @minimum 0
     * @maximum 17
     */
  row?: number;
  /**
     * @minimum 0
     * @maximum 7200000
     */
  atMs: number;
}