# lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RunAction } from './runAction';

export interface ScoreInput {
  /**
     * @minLength 1
     * @maxLength 10
     * @pattern ^[A-Za-z0-9_-]+$
     */
  nickname: string;
  /** @minimum 0 */
  score: number;
  /** @minLength 1 */
  proof: string;
  /** @maxItems 1000 */
  actions: RunAction[];
  /**
     * @minimum 0
     * @maximum 7200000
     */
  endedAtMs: number;
}